requirements.txt
torch
matplotlib
numpy
Pillow
scikit-learn
